+++ plugin/results.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 import random, logging
 from flogin import ExecuteResponse, Result
-from .library import SphinxLibrary
+from .library import Library
 import webbrowser, asyncio
 
 if TYPE_CHECKING:
@@ -14,7 +14,7 @@ log = logging.getLogger(__name__)
 
 class ReloadCacheResult(Result["RtfmPlugin"]):
     def __init__(self) -> None:
-        super().__init__("Reload cache", icon="assets/app.png")
+        super().__init__("Reload cache", icon="assets/app.png", auto_complete_text="".join(random.choices("qwertyuiopasdfghjklzxcvbnm")))
 
     async def callback(self):
         assert self.plugin
@@ -33,7 +33,7 @@ class ReloadCacheResult(Result["RtfmPlugin"]):
 class OpenSettingsResult(Result["RtfmPlugin"]):
     def __init__(self) -> None:
         super().__init__(
-            "Open Settings", icon="assets/app.png", sub="Open the settings webserver"
+            "Open Settings", icon="assets/app.png", sub="Open the settings webserver",auto_complete_text="".join(random.choices("qwertyuiopasdfghjklzxcvbnm"))
         )
 
     async def callback(self):
@@ -48,7 +48,7 @@ class OpenSettingsResult(Result["RtfmPlugin"]):
 class OpenLogFileResult(Result["RtfmPlugin"]):
     def __init__(self) -> None:
         super().__init__(
-            "Open Log File", icon="assets/app.png", sub="Opens up the flogin log file"
+            "Open Log File", icon="assets/app.png", sub="Opens up the flogin log file",auto_complete_text="".join(random.choices("qwertyuiopasdfghjklzxcvbnm"))
         )
 
     async def callback(self):
@@ -62,7 +62,7 @@ class OpenLogFileResult(Result["RtfmPlugin"]):
 
 class OpenRtfmResult(Result["RtfmPlugin"]):
     def __init__(
-        self, *, library: SphinxLibrary, url: str, text: str, score: int
+        self, *, library: Library, url: str, text: str, score: int
     ) -> None:
         self.library = library
         self.url = url
