+++ requirements.txt
@@ -1,4 +1,4 @@
-flogin==1.0.0
+git+https://github.com/cibere/flogin
 favicon==0.7.0
 Wand==0.6.13
 aiohttp==3.11.10
